Add files via upload
By creating three files (HTML file, one CSS file and JavaScript file), I also have an image of the clock that I used in the background and in addition to that, I made one hour, minute and second hand (HTML and CSS usage). These hands rotate according to the time of the system (I used the predefined date function of JavaScript to calculate the degree of rotation of each hand).